refactor: move device compatibility check to utils and update SmashConfig
* change: relocate check_device_compatibility function to utils.py for better modularity
* update: modify SmashConfig to utilize the new utility function for device validation
* enhance: improve code readability and maintainability by removing the static method from SmashConfig
* add: introduce unit tests for device compatibility checks in test_device.py
